refactor to fix type warnings

+++ Bunny.cpp
@@ -3,11 +3,11 @@
 Bunny::Bunny(int textureIndex, ZL_Rect &bounds)
 {
 	nSpriteIndex = textureIndex;
-	dGravity = 0.75;
-	dx = 0.0;
+	dGravity = 0.75f;
+	dx = 0.0f;
 	dy = ZLHEIGHT;
-	dSpeedX = ZL_Rand::Range(0.0, 10.0);
-	dSpeedY = ZL_Rand::Range(0.0, 10.0) - 5.0;
+	dSpeedX = ZL_Rand::Range(0.0f, 10.0f);
+	dSpeedY = ZL_Rand::Range(0.0f, 10.0f) - 5.0f;
 	rectBounds = bounds;
 }
 
@@ -19,27 +19,27 @@ void Bunny::Update()
 
 	if (dx > rectBounds.right)
 	{
-		dSpeedX *= -1;
-		dx = rectBounds.right;
+		dSpeedX *= -1.f;
+		dx = (float)rectBounds.right;
 	}
 	else if (dx < rectBounds.left)
 	{
-		dSpeedX *= -1;
-		dx = rectBounds.left;
+		dSpeedX *= -1.f;
+		dx = (float)rectBounds.left;
 	}
 
 	if (dy < rectBounds.bottom)
 	{
-		dSpeedY *= -0.85;
-		dy = rectBounds.bottom;
-		if (ZL_Rand::Range(0.0, 1.0) > 0.5)
+		dSpeedY *= -0.85f;
+		dy = (float)rectBounds.bottom;
+		if (ZL_Rand::Range(0.0f, 1.0f) > 0.5f)
 		{
-			dSpeedY -= ZL_Rand::Range(0.0, 6.0);
+			dSpeedY -= ZL_Rand::Range(0.0f, 6.0f);
 		}
 	}
 	else if (dy > rectBounds.top)
 	{
 		dSpeedY = 0;
-		dy = rectBounds.top;
+		dy = (float)rectBounds.top;
 	}
 }
